refactor(inbox): extract InboxPage data, selection, plan-summary, status, and top-bar layers
InboxPage was 1012 lines with a cognitive complexity of 30 against a
ceiling of 20, ~78 hook calls, and a ~450-line render block. Movement
only: no behaviour, prop value, or data-testid string changes.

- useInboxListData: list/open-plan queries, root derivation, rescan
  targets, destination roots, and the page-level search filter.
- useInboxSelection: URL-backed selection plus the reclassify_v2 re-split
  and classify-split handoffs and their shared stale-cleanup gate.
- useInboxPlanSummaries: per-ingestion frame-type hint and the merged
  three-source breakdown tallies PlanPanel renders.
- useInboxPageStatus: derived stats, folder/master summary, and the
  memoised status-bar node.
- InboxTopBar: FilterToolbar wiring and the review-plans / bulk-confirm /
  rescan actions, carrying their render conditionals.

The InboxDetail remount key stays inline in InboxPage: InboxDetailProps
is already exported, so a wrapper component would have forwarded twelve
identical props for one key expression.

FindSpotlight.anchors.test.ts follows the inbox.confirm-row anchor to its
new file. The anchor string itself is unchanged.

Refs: astro-plan-kyo7.112
